CR: rename module emailOnCompletion » job_email_report
I think that the new name describes better what the module does instead of
describing how it was used by its consumer.
