+++ src/MCPClient/lib/clientScripts/emailFailReport.py
@@ -34,7 +34,7 @@ from main.models import Job, Report
 from custom_handlers import get_script_logger
 from externals.HTML import HTML
 
-from lib import emailOnCompletion
+from lib.job_email_report import run_job
 
 
 django.setup()
@@ -211,4 +211,4 @@ if __name__ == '__main__':
     content = get_content_for(args.unit_type, args.unit_name, args.unit_uuid, html=False)
     store_report(content, args.unit_type, args.unit_name, args.unit_uuid)
 
-    emailOnCompletion.run_job(args.unit_uuid, False)
+    run_job(args.unit_uuid, False)
