+++ src/MCPClient/lib/clientScripts/post_store_aip_hook.py
@@ -15,7 +15,7 @@ from main import models
 # archivematicaCommon
 from custom_handlers import get_script_logger
 import elasticSearchFunctions
-from lib import emailOnCompletion
+from lib.email_job_report import run_job
 import storageService as storage_service
 
 logger = get_script_logger("archivematica.mcp.client.post_store_aip_hook")
@@ -125,7 +125,7 @@ def post_store_hook(sip_uuid):
     # DSPACE HANDLE TO ARCHIVESSPACE
     dspace_handle_to_archivesspace(sip_uuid)
 
-    emailOnCompletion.run_job(sip_uuid, False)
+    run_job(sip_uuid, False)
 
     # POST-STORE CALLBACK
     storage_service.post_store_aip_callback(sip_uuid)
